Change click behavior for open and selected links
Previously, clicking an open link would toggle it to closed.

However, it seems more likely that the intention of a person clicking on
an open but unselected link is to select it, so now that is the behavior
for clicking local and global links that are open.

By contrast, clicking on a selected link has no purpose except to close
the link, which now is what happens.