Add full support for question mark topics
Previously, question marks in topics would be allowed in the header
string but would get removed from other parts of the app like the URL
bar, which made sense because it would probably get interpreted as a
query string.

This commit makes question mark topics a first-class topic that can
exist everywhere in the app, including JSON filenames and URL paths.
The problem of question marks getting parsed as query strings is
handled by changing how we access the path when ascertaining the
current URL, previously it was using the location.hash lookup which
ignored query strings, but now we are just taking a simple substring of
the URL after the origin which includes question marks.

This commit also introduces a topic name permutation called
requestFileName. Previously we were URL encoding filenames and URL
encoding the HTTP request paths, and the server was decoding the URL
and expecting file names that were URL decoded, so now we only encode
the URL string and the server decodes it before looking for static
assets by that name.

Question marks are removed from topics at the "caps" stage because two
topics should not differ merely on a question mark, and by that logic
the caps stage should remove quotation marks also.