Support colons in topic names
Previously, the definition of a key in the parser prevented escaping
colons (or commas, periods, and semicolons) because any colon would be
considered the end of the key.

Now, we ignore escaped colons, producing a topic with back slashes,
which should be removed by the first pass of the Topic class in creating
the display string.