Support import self-references
Previously several limitations technically prevented creating an import
reference to a local link of the current topic.

First, we needed to make it so that local links that are redundant get
checked for possible import references, or have provisional local links
converted to import references if they later become redundant. This is
necessary just for the import reference not to be classified as a
redundant local reference in the first place.

In this commit, we had to tell the local reference matcher to reject
as candidate local references a local reference to the topic subtopic,
because even though this is technically a subtopic of the given topic in
certain respects, there should never be a local reference to it because
it is displayed by default as the root of the topic tree, and its parent
link is either a global reference in the lowest paragraph of the prior
topic tree, or if this is the root topic tree there will be no parent
link. Therefore any reference that is to the current topic's root topic
is not a valid local reference in principle. Having the local reference
matcher ignore these cases causes them to be forwarded to the import
reference matcher which is lower in the matcher list.

Secondly, this commit fixes a bug wherein Paragraph#path identified
a change of topic tree for path purposes by climbing the DOM from
section to section and waiting for the section topic to change,
however, this doesn't work if two adjacent topic trees share the same
topic, so now we check whether the current section element's topic
DOM node is the same as the parent element's topic DOM node, if so,
then they shouldn't both get path segments, if not, then there should be
two adjacent path segments with the same topic.

This commit also retroactively adds checks to previous browser tests
that expected elements are visible.