Up on an import reference path should select parent import reference
Previously, down on an import reference selected the target path, and
then pressing up selected the local parent, whereas the user might
expect symmetry. The downside of this change is that if a given path has
only one link per level, there is no way to get to the local parent of
the import path. Hopefully if it is a sparse tree, there isn't much
reason to navigate around.