Auto-scroll to position on initial page-load
Previously, if a link was pre-selected on initial page load, we would
scroll smoothly to it, which was odd because the page loaded already
scrolling eg if reloaded. Now, we check if this is the initial page
load, and if so, we snap to the right position immediately, rather than
doing a smooth scroll.