Test "tryPathPrefix" behavior
If someone goes to a path like A/B/C that is invalid because C does not
exist, or because there is no valid link from B to C, the system should
fail gracefully, trying A/B and A before throwing an error.

What happens when B doesn't have a link to C? The render code recognizes
it during the initial recursive descent, and aborts the render of C, but
does not error, so that when the displayPath function goes to retrieve
C, it finds that it isn't there, and retries the subpaths. If we would
throw the error during the render process, then we wouldn't finish the
render process in order to be able to do the retries without error
later during the display process.