Remove concept of blocks from parser
Previously the data model was that a JSON file had a topic key, and then
a list of subtopic keys, and each subtopic key had one or more "blocks",
which could be of type text or table or list etc, and then within the
blocks there were tokens like local and global links, etc.

However, this added complexity because there were block renderers and
also token renderers which were very similar but in two places in the
code, and it made reading the parser or renderer more difficult because
one had to burrow into the block handler code and then into the token
handler code.

This commit replace the concept of blocks with tokens, so now,
multi-line entities can also be tokens, and they can contain tokens,
and the rendering code has consolidated the rendering of blocks and
the rendering of tokens into one recursive renderTokenElement function
that calls itself if one type of token can contain tokens of a different
type. This also simplifies the JSON output and makes it more readable,
because most of the time the paragraph is just a single text block, and
so it was a lot of boiler plate to have an array of blocks with one item
that had a tokens array instead of just making the subtopic key have a
single tokens array for the whole paragraph.

The parser passes around a tokens array and the subroutines add tokens
to it. This seemed preferable to having pure functions return tokens
because anyway the subroutines are a function of the existing tokens.
However, it seemed preferable that parseText be a pure function,
however, previously local reference conversion worked by storing a
reference to a token array and later splicing the array and replacing
the local reference token with an import reference, however, that means
parseText needs a handle on the final array that will go into the JSON
object such that it can splice it, so parseText needed to be given the
array which it would then return, which was a bit counter-intuitive and
brittle because changing this would silently break local reference
conversion, so now we convert the actual local reference token object
itself, which avoids needing to have a handle on the array that will
end up being in the JSON.