Allow escaped style characters in topic names and fix escaping in text
Previously, backslashes were only permitted in the text body, in order
to eg cancel a double square brackets being interpreted as a link, or
to cancel an underscore being interpreted as italics, etc.

However, due to a bug in renderStyledText, this was only ever working
because the backslash caused multiple text tokens to be created which
were rendered individually, so escaped underscores were never seen as
matching sets, but if they would be in the same token, they would have
been.

Additionally, backslashes were being removed from topic names in the
topic class so they never reached the browser.

Now, we keep backslashes all the way until we get to the DOM, and now
renderStyledText handles them properly, removing the backslash and
not interpreting the subsequent character as a style character.

Additionally, this commit simplifies the code of consolidateTextTokens
which seems to fix an error that was preventing adjacent text tokens
from being consolidated, which indirectly caused escaping of style
characters to work.

This commit also introduces a new constructor argument to the Topic
class that is used when the string being passed in is already the
mixed-case string, to avoid double-processing, which is now relevant
because we remove backslashes when creating the mixed-case string for
use eg in the URL, and so because path strings are parsed to retrieve
the path array, if the topic "\_Hello World\_" gets turned into a mixed-
case string, we will note the backslashes and not remove the style
characters correctly producing the user-facing "_Hello World_", but
when that value is parsed a second time, this time the underscores do
look like style characters and they are removed, so this commit
introduces a boolean that lets you indicate at Topic instantiation time
that the string being entered is _already_ in mixed case, to avoid this
accidental double-processing now that mixed-case string generation is
not idempotent.

Finally, some characters like "_" when encoded eg to %5f, if used in a
URL will be converted back to "_" for display purposes, however, this
defeats our purpose for encoding these values, that the path string
should be reversible to retrieve the mixed-case path-array. So, the
character "_" and in future any other such characters are now double
encoded, and the path array strings are double decoded before
instantiating the Topic class. Perhaps double encoding prevents a person
from using the single encoding as a meaningful topic name, but it seems
unlikely that people would ever use a URL encoded character not
intending it to be decoded.

This commit adds a browser test that escaped style characters in topic
names works, and also a small assertion that they work in regular text
also, pending more complete style browser tests.