Allow canopy to be run with nodemon
When developing the front-end code, the `canopy watch` command allowed
webpack's new build to be copied to the current project's build
directory so that on refresh the new code would be visible in the
browser.

However, if changes were made to the parser code itself, these changes
were not reflected in the `canopy watch` command because the old parser
code was still in memory.

This commit introduces the ability for an optional development mode
wherein the developer can run `canopyd` instead of `canopy` and it will
run with nodemon and refresh the `canopy watch` command on code changes.

In order to achieve this, the canopy binary has been converted into a
canopy.js file, which either node or nodemon can run.

This commit adds a nodemon.json config file, but specifies the watch
path in the binary so as to use the CANOPY_LOCATION variable.

This commit also replaces the newBuildDirectory canopy build option
with its inverse, keepBuildDirectory, because this seems like the more
expected behavior, especially in development when incorrect build files
can accumulate.