Add browser tests for style markup
This commit introduces browser tests for style markup like links and
images, and fixes many outstanding bugs in that functionality.

Escaped characters now do not create a new text token but rather
augment the existing token, in order to allow style characters to span
escape characters.

This commit adds more steps to the parsing of code blocks and style
characters to ensure that no text within code blocks has special
meaning, and to make sure that style characters are parsed correctly,
eg only if they are opened and closed, and in the right sequence.

This commit introduces a .canopy-image class to limit Canopy css only
to markup images and not raw html.

Additionally, this commit makes slight changes to the logging output
of the build command.