Handle topic names with % & + $ and ()
This commit makes several changes to the existing system of topic
name handling.

Previously, we wanted URLs to be reversible to retrieve the mixed-case
topic name. However, if the topic name had underscore literals in it,
then when spaces were converted to underscores for the URL slug, this
irreversibly mixed up real underscores with former-spaces. Initially
we were double-encoding underscores and double-decoding the URL, but
this presented challenges for supporting real percent signs which were
not part of an encoding, or topic names that contained encodings meant
literally.

So, this commit uses a new approach wherein percent signs in the real
topic name are encoded, characters like quotes and ampersands that do
not break the app if unencoded are left unencoded for readability, and
characters like # that do break the app are encoded, and then at page-
load, we take the URL which is now lossy from having converted spaces
to underscores along with real literal underscores, and we convert all
underscores to spaces and URL decode the string. In order to handle the
lossy underscores, now Topic#caps converts all underscores to spaces,
and topic names are now not permitted to differ on a character that is
an underscore in one and a space in the other. This requires making
changes to places in the app where topic strings are compared,
previously we had been using Topic#mixedCase for readability, but now
all comparisons must use Topic#caps to ensure lossy conversions still
find the original topic.