Handle invalid single-segment paths
Previously, the display code contained the logic for trying path
prefixes until none remained and then navigating to the default topic.
This was enabled by having the render code fail silently if it didn't
find a given topic, so that whatever portion of the path is valid gets
rendered, allowing the display code to recursively try path prefixes
until one is extant.

This worked most of the time because path rendering is advanced by
global references, so calling `fetchAndRenderPath` for the path
`A/B/C` will only request `C.json` if B has a valid global reference to
C. If it doesn't, then `fetchAndRenderPath` never gets recursively
called with the path `/C`. Only when some portion of the DOM path
already exists do we call `fetchAndRenderPath` from
`fetchAndRenderPath`, but then also we check `Path.connectingLinkValid`,
so it was never possible to make a failing request previously unless
the invalid topic was the very first topic, which wasn't tested.
However, it was therefore possible to have a failing network request in
the render code.

If the render code had a failing network request, this was a problem
because previously, the call to displayPath was in a `then` call, so
it only ran if the request was successful, whereas we want it to run
in all cases because that is where the tryPathPrefix logic is, so now,
the display code is in a `finally` block, which will run regardless of
whether the network call was successful.

This code also removes failed requests from the request cache in case
it was a temporary failure, so that the request will get tried again
on the following user interaction.