Use display topic names everywhere possible
Previously, we only used the display version of the topic name for the
page header, and for all logical comparisons we used the mixed-case
version, and it was the mixed-case version that was given in tokens
from the server to the client. However, this meant that we were
initializing Topic objects from the mixed-case string, which required
the creation of a constructor option to do so, in order to avoid
double-processing. However, it seemed as though this was a dangerous
direction, because the more we are instantiating Topic objects with
strings that were not the original, the more chance there was of
unintentional double-processing, eg in the case of underscore literals
which if double-processed are no longer escaped and are interpreted as
style characters and removed to produce the mixed-case string.

It was not possible to remove the fromMixedCase constructor argument
entirely because in one place we take the path from the URL and
convert it into topic objects. This is undesirable because the path
isn't even really mixed-case, because once you convert spaces to
underscores, converting them back produces a corrupt string that has
lost any original underscore literals. However, because the path topics
are only used for caps, it happens to not matter that they were
instantiated from a corrupted string. So, for clarity the Topic
constructor argument has been renamed fromSlug to indicate that when
instantiated from a converted path string, it wasn't even instantiated
from the mixed-case version, and so now, such a topic has a null display
and a null mixed-case, to indicate that only its slug version and
derivatives are reliable, because these are lossy on underscores the
same way the original path string itself is, so there is no worry of
trying to reverse the lossy slugification process and getting corrupt
data.