Test # and path prefix mode
This commit adds tests for the hash-urls option and the
--project-path-prefix option. These are build options, so we have to
create new project directories/servers for them running on different
ports in order to test this.

This commit fixes a problem in the hash URL feature that was introduced
when we changed how we captured the window location, and so were no
longer removing the hash signs properly.