Allow arbitrary newlines between paragraphs in expl files
Previously, we were splitting on the literal `\n\n`, so if there was
an odd number of newlines, the paragraph would begin with `\n` and it
would register as a block with no key. Now, in the case where the first
character is `\n`, we slice the newline and increment the line number,
and then the key has the correct line number in the subtopic line
numbers store in parserContext, and then the parsing code can just
trim all paragraphs knowing that once it finds the right key, it will
look up the correct line number for that subtopic in parserContext.

We only need to worry about one newline because if there is an even
number, all the newlines will be handled as empty-string paragraphs
and wont obscure any real paragraphs, and in the case of an odd number
we only need to worry about the one odd newline.