Use prior link selection in key browsing
Previously, pressing down on a parent link selected the first child
link no matter what.

Now, we keep track of what link was last selected per paragraph, and so
when we need a generic link from that paragraph eg upon pressing down
on a parent link, we select the previous one and not the first one.

This also resolves a long-standing issue wherein pressing down on an
import reference and then pressing up repeatedly returned one to the
import reference's global reference, not the import reference itself,
which was unsatisfying. (Perhaps if one enters an import reference and
then navigates to a sibling and then presses up, one would be expecting
to return to the actual parent link, but this oddity would only happen
if one had arrived at a child paragraph from entering an import
reference, so it seems not too confusing.)

Now that we preserve last link selection per paragraph, why do we also
need to preserve last link selection per URL? In eg a paragraph of URL
links, indeed either store would give you the same answer. However,
in the case of an import reference, now that the following commit
causes the import reference and the "imported" local reference to have
the same URL when selected, and so knowing which selection state a
given URL has would not be stored in the "most recent link selection of
paragraph" store, because both links might be the most recently selected
of their respective paragraphs, but we would know from the 
session/history stores which of the two links were most recently
selected from that URL. Really the necessity of these stores decreased
when most link selection states got unique URLs, removing the need to
store this data in the session.