Selected import references should have URL path of target paragraph
When a global link to B is selected in paragraph A, the URL path is
A/B. When a local link is selected in paragraph A to subtopic B, the
path is A#B. In both cases, selecting a link makes the URL represent
the path of the paragraph that is previewed given the link selection.

However, until now this was not so for import references, because unlike
with global and local links, the target paragraph path for the import
reference is already the URL path for the link selection state where
the target imported link is selected. (For globals, there is no such
thing as selecting the root paragraph, and so A/B doesn't compete with
the selection state for any child parent link, and for local links,
it is clear that URLs and selected link states map 1:1, so it is only
import references where one URL could logically describe two selected
link states.)

However, this behavior seemed confusing because every other link
selection corresponds to the child paragraph path, whereas for import
references the URL path would be for the parent paragraph, which was
unrelated.

Now, both link selection states, selecting the import reference, and
selecting the imported reference, both have the same URL, eg
A/B#C. Because link selection history is done by URL, that means that
on refresh or new tab, the URL corresponds to which ever link selection
was most recent.

On opening the URL with no localStorage/history, the child link
selection wins out because looking at the URL, A/B#C most directly seems
to concern the paragraph A/B#C, and so it makes sense that the selected
link should be its immediate parent, not the import reference.