Change literal underscores in topic names to encoded escapes
This commit deals with the ongoing question of how to turn mixed-case
topic names into a URL path that is then usable to find the right topic
paths to HTTP request, and the right keys to find the subtopics in the
response JSON.

Initially, the "slug" was just the mixed-case version of the string
with underscores instead of spaces. This didn't collide with markup
underscores because already markup underscores were being removed
entirely from mixed-case and therefore slug. However, with support for
escaped underscore literals it became clear that the conversion of
mixed-case to slug was lossy because spaces converted to underscores
became indistinguishable from the underscore literals that were already
in the mixed-case string. (It doesn't help that underscore literals
had to be escaped by backslashes because these also were removed once
style characters were removed to produce the mixed-case version and
there was no longer possibility of confusion between style underscores
and underscore literals.)

So, initially the solution was to double encode underscores to %255f
because a single encoding got converted to a regular underscore in
Chrome. However, this collided with the literal %5f that the user might
have entered which also encodes to %255f.

So the design was changed to make the path slug string
an irreversably lossy version of the mixed-case string, and that in
order to parse the slug and request the correct JSON files and to
find the correct subtopic keys in those JSON files, all comparisons and
requests were done with the caps version, which removed all spaces,
thereby making it irrelevant that literal underscores and spaces turned
into underscores for the slug had collided and were all reversed to
spaces, because now all spaces were removed anyway. However, this
approach required putting multiple versions of the topic and subtopic
string in the DOM datasets, and meant that the string that was in the
link token target was not the same as the subtopic key in the JSON file,
which meant that any consumer of the JSON would have to know about the
different string conversions which really were only necessary due to a
browser implementation concern of DOM string limitations, and shouldn't
have to apply in the case of eg a command line browser application.

So, the decision was made to return to an approach wherein the URL
slug was reversable to retrieve the mixed-case string, which would be
used in most comparisons and identifications, eg in the subtopic keys
of the JSON file.

The problem remained that in a world with underscore literals in topic
names, and where spaces are converted to underscores, literal
underscores needed some encoding that differentiates them from spaces
converted to underscores that will need to be converted back. In theory
we could use the URL encoded underscore %5f, but this is not generally
seen as part of the spec and Google Chrome for example will convert
"%5f"s back to underscores, colliding with spaces-turned-into-
underscores. Initially we had double encoded these literal underscores
to %255f, but then it was realized that this collided with a user who
was using as a topic name the literal %5f. So it was decided that just
as backslashes represent an escape character in the expl file itself,
the encoded escape character %5C before an underscore can represent an
"escaped" underscore, and %5C%5C can represent an escaped backslash
itself, much in the style of most programming languages and text
formats. This gives us a way to reversably encode characters that do not
have URL encodings, while at the same time having no collisions with
other text input, as URL encoded characters will always contain a %
which if it had been in the user-input would get encoded to %25, or
a backslash which if it would have been in user-input would have gotten
encoded to %5C%5C, and so because encoding expands these special
characters to double/longer forms, when we find them in the URL in the
singular, ie either % or %5C as opposed to %25 or %5C%5C, we can be sure
that they were put there by the encoding scheme and not by the user,
and so we can confidently URL decode them and in the case of single
backslashes see them as a temporary escaping character rather than real
data.

This commit introduces CSS escaping to the Topic#escapedMixedCase
definition to allow backslashes in topic names ie data attributes.

This commit introduces different methods of instantiating Topics to
make it clearer what types of input there are and what transformations
each source requires.