Reorganize error handling in updateView
Previously, updateView would render the page, and then it would display
the correct path, and then it would catch errors, but all the while
certain parts of the display code call console.error rather than
throwing, and so if an HTTP request failed, you would end up getting
the display errors in the console before the actual HTTP error, which
only was logged upon the later `.catch` call, so now the render code
has a catch that logs errors, followed by the display logic which is
caught to handle errors.

In addition, this commit now has errors thrown as Error objects, which
retain certain helpful metadata.