Redirection to a different root topic should not animate scroll
When pressing browser back and going to a new page, the concept of
scrolling isn't really appropriate because you are going to a new page,
not a new position of the current page, and so now we just snap the page
to the old position without animation.

This commit also introduces scrolling when there is no link selected,
now we scroll without animation to the root paragraph.

Additionally, now scrolling is based on the location of the selected
link, not the selected link's enclosing paragraph.

Other places in the app where a redirect is done to a new page,
the redirect is not animated, whereas scrolling within a given page
including using the browser back functionality does animate.

This is all enabled by a new convention where displayOptions can contain
a scroll style that determines whether there should be animation.