Front-load image tokens
At the moment, images are styled as being inline float right, and so
it seems we would always want them top-aligned, but this only happens
if they are before the other elements in the given container, so for now
we're just going to front-load the image tokens to the beginning of the
paragraph token array, although we may revisit this if people want more
fine-grained control over where images show up in the text. Our
assumption is that paragraphs should really only have one image, as
they should be quite small, and so they should almost always be aligned
with the top right corner of the paragraph, which if they are an inline
floated element means they should be the first element.