Change tab-opening shortcuts
Previously, it was unclear what meta/control-return should do, ie
whether the new tab should be at the existing path, or whether it should
be "zoomed" to the new path of the given link. Because there was no way
to open a new tab at the current path from the keyboard, it was decided
that that was what it should do. However, that wasn't in keeping with
the general pattern of the return key having "zoom" properties.

This commit introduces a new shortcut, "d" for duplicate, which opens
a new tab at the current path. This opens up meta/control return to
zoom, in keeping with return as a zoom key.

However, plain return on a local link seems like it should be a synonym
for "down", and the idea of alt-return sounds redundant given both of
those keys have an association with zooming. So return on a local link
will just advance due to the exceptional nature of local references,
which have no connotation of pointing somewhere else as do global and
import references, and alt-return isn't necessary because the "z" zoom
command does the same thing and is simpler to understand. The downside
of this decision is that plain return on a local and meta-return on a
local do not do the same thing.

So in summary:

Return on global/import - redirect
Return on local - advance
Control/Meta return on global/import - redirect
Control/Meta return on local - zoom

Control/Meta click on global/import - inline at current path
Control/Meta click on local - inline at current path
Control/Meta+Alt click on global/import - redirect
Control/Meta+Alt click on local - zoom

d - duplicate