Catch "link selector null" error in display code
Callers of updateView can specify what link they want to have selected,
and one of the options for specifying a link is a callback that should
return a link object if invoked after page render.

However, sometimes this callback does not return a link. In this case,
the page should just render without a link selector argument. However,
how best to recognize this case and change tack? Two options previously
employed were one, throwing an error that was caught in updateView,
although this cluttered the main function with a very marginal case,
or, links could fail silently and respond to a Link#empty method, but
this required all users of the link class remembering to check if the
link was empty. This new option lets us use the explicitness of an error
but handles it close to where the logic is relevant, in the display
function where we are taking the linkToSelect argument in the first
place and doing things with it.