Initialize Canopy Sketch framework
Canopy sketch will be an interactive CLI that allows one to build
content from the bottom up, beginning before it is clear what the
top-level topics and references of the project are.

The general architecture will be a map of state objects plus input
characters to resulting state objects, and rendering will be a function
of state object.