Make sketch framework pure DFA
Previously, there were cases like multi-key input where the CLI wasn't
a pure DFA. Now, complex behavior like multi-key input or yes/no
lists, or even alerts that return you to a given caller state are
not native behavior, but are built out of a series of states and single
key transitions. Helper functions will assist in creating groups of
changes to the Modes object that coordinate to produce a given feature.