Add `canopy bulk --sync` command and improve other logging
Now running `canopy bulk --sync` will generate a `canopy_bulk_file`
that can be edited repeatedly to rebuild the topic directory and
by extension the JSON build files.

This commit also improves the log messages of the watch and new sync
command to include simpler time stamps and a mention of the specific
file changed, and adds colors to other logging.