Change default topic file to path and add validation
Previously, the `.canopy_default_topic` file held the key of the
default topic, rather than the path to the file, but then checking
whether that topic existed was more difficult, and so now we store the
path, and throw an error if the default topic path doesn't correspond
to a real file. In bulk mode, we validate on ending a session and log an
error if there is no longer a valid default topic file.

The `.canopy_default_topic` file is now just `canopy_default_topic`
and not a dotfile, because the user should be aware of it.