Ignore category notes file keys if not matching file name
Sometimes a category notes file will by chance have a first note that
begins with a question, which gets incorrectly labeled as a topic.

Now, if a paragraph is in a file that is a category notes file, eg
`A/B/C/C.expl`, then only a root key of "C" will be interpreted as a
topic, and all others will be interpreted as notes and ignored by the
parser.