Remove SVG for URL links and open in new tab
Previously users were warned about external URL links with an external
URL svg, but this caused problems because it could render on a new line
after the given link text which was confusing. Now, we remove the svg,
and avoid users confusing URL links with regular links by making them
open in a new tab always, to lower the consequences of a mistaken
trigger.