Allow keys without values in `expl` files
Previously, a key without a value would be considered a note. Eg,

```
Hello world:
```

would be a note and not a paragraph node. However, when sketching out
a topic tree, it is helpful to be able to create placeholder nodes so
that one can create references to them without producing runtime errors.