Change error format to [Subtopic (Topic)]
Previously, errors would speak about a problem existing in
[topic, subtopic], but this format was a bit confusing to read, and if
it was the topic subtopic then it would be duplicative, whereas now
we say either [Topic] for the topic subtopic or [Subtopic (Topic)] for
a subtopic, which is easier to read. This commit also fixes problems
with tests that were unnoticed due to a mistaken .only call.