Bug: Question mark keys should not include question marks
Previously the key ABC? DEF? would match the second "?", when really
"ABC?" should be the key, as keys are not permitted to contain "? ".

A key should not be able to include a colon or question mark, or any
other sentence-terminating punctuation.

This commit also now allows semicolons and commas in topic names, and
allows topic names like "Canopy.js" that include sentence-terminal
punctuation, but not followed by a space.