Return on global should not select link
Previously, return redirected to the global target as the root of a
new page, but it also selected the first link, because of a general
tendency to maintain a link selection once the user has selected a link
initially. However, this causes too much of a jump visually and in the
URL, going from /A/B to /B/C in one jump. So now, the return key just
takes us from A/B to /B, and the user's first keypress selects the first
link.