Bug: Ensure that file paths exist before reading them
If we load a set of files and then delete one independently, then
when we go to read the disks from file as part of the bulk file
generation process, we will try to read a non-existent file, so now
we check that the files exist before reading them.

It isn't a problem for a file that was in the original selection not to
be in the new selection because the original selection in the first
place is just a check to make sure we only load files selected by the
user, but if the user has deleted a file that was selected, they have
expressed a disinterest in that file, and so it is fine to not include
the file in the next generated bulk file.