Alternate block quotes based on auto-detected direction
If a block quote is of a language that is RTL, the block quote border
line should be on the right-hand side, so this code uses a hack to
have the browser assign an auto direction and then detects that setting
and applies the direction property explicitly so that the css
can differ based on text direction. This commit also makes block quotes
handle RTL text in the first place.