Handle non-expl files in the root topics directory
When generating a file set for all on disk files, we include non-expl
files because we need to know about them eg in case their presence
should cancel the deletion of an otherwise empty category directory.
However, they're not going to have a valid category path, because
they're in the root topics directory. So now we return null explicitly
for displayPath to avoid a regex error, and add a new condition that
for a directory to be rendered in the bulk file, it must have at least
one .expl file.

If there is a valid .expl file in the root topics directory, the bulk
mode will render the category header as [PLEASE SET CATEGORY] even
though technically the parser can handle expl files in the root
topics directory.