Change images to be full-width
Previously, images created using the markdown syntax would be rendered
on the right side of the text of a paragraph. However, this caused many
problems, as many paragraphs do not have enough text to fill to the
height of the image, making the image taller than the text creating a
lop-sided visual.

In addition, the entire convention of images on the right side of text
articles is an homage to the idea of linear documents, where a single
static resource must accomplish multiple goals because only one linear
experience can be transmitted. Canopy, by contrast, encourages
customized assembly of single-purpose resources, ie, rather than making
one composite experience (large text plus images) for everybody, Canopy
encourages the author to make content a la carte, organizing text as
narrow explanations of specific things, and labeling images as being a
separate optional resources with separate identifiers, so that if the
user requests an explanation, they get narrow text, and the text might
make mention of an available image, and the user can request that image
as a second follow-up request for a second resource. This approach
allows us to have the state of the page match the focus of the reader,
ie, rather than giving the reader a composite document and having them
sometimes looking to the left and sometimes looking to the right, in
Canopy every URL and page state has one centered visual, and if the
user is supposed to look at something else, that other thing has a
different URL and page state. Thus, navigating to a given page in
Canopy guarantees the user will see the single point being made on that
page in a much more fool-proof way than if the reader is offered
composite resources with many points and aspects per unit, some of
which may be missed.